requirements.txt
# Runtime dependencies - pinned to minor versions for stability
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
gunicorn>=21.2.0,<21.3.0
pydantic>=2.5.0,<2.6.0
pandas>=2.1.0,<2.2.0
numpy>=1.24.3,<1.25.0
pytz>=2023.3,<2024.0
matplotlib>=3.7.2,<3.8.0
reportlab>=4.0.4,<4.1.0
jsonschema>=4.19.0,<4.20.0
qrcode[pil]>=7.4.2,<7.5.0
python-multipart>=0.0.6,<0.1.0
jinja2>=3.1.2,<3.2.0
aiofiles>=23.2.1,<24.0.0
typer>=0.9.0,<0.10.0
slowapi>=0.1.9,<0.2.0
python-magic>=0.4.27,<0.5.0
tzdata>=2023.3

# M12 Compliance dependencies for PDF/A-3 + RFC 3161
rfc3161ng>=2.1.3,<2.2.0
PyPDF2>=3.0.1,<3.1.0
cryptography>=41.0.7,<42.0.0
lxml>=6.0.0,<6.1.0

# M13 Auth dependencies for JWT and email
PyJWT>=2.8.0,<3.0.0
email-validator>=2.0.0,<3.0.0

# Payment processing dependencies
stripe>=7.0.0,<8.0.0

# Database dependencies
sqlmodel>=0.0.14,<0.1.0
asyncpg>=0.29.0,<0.30.0
boto3>=1.34.0,<1.35.0

# Testing and development dependencies
httpx>=0.28.1,<0.29.0
pytest-asyncio>=0.21.0,<0.22.0